[MMG_RPG] TCP 서버에서 NPC 데이터 캐싱해오기

+++ MMG_AdminTool/MMG_AdminTool/Models/NpcSpawnViewModel.cs
@@ -6,9 +6,9 @@
         public int TemplateId { get; set; }
         public int MapId { get; set; }
 
-        public float PosX { get; set; }
-        public float PosY { get; set; }
-        public float PosZ { get; set; }
-        public float DirY { get; set; }
+        public int PosX { get; set; }
+        public int PosY { get; set; }
+        public int PosZ { get; set; }
+        public int DirY { get; set; }
     }
 }
